Add files via upload
The Student Information system is a web based application about college, Nepal Commerce Campus. From this application, admin can store, update  and delete the students information and also provide the information about the events and some notice for the users by an admin. This project is developed on PHP (Laravel 5.3)